requirements.txt
fastapi
uvicorn
pydantic
MetaTrader5
ta
httpx
requests
python-dotenv
aiofiles
